move class TableNames to oshdb-util module
