+++ oshdb-api/src/main/java/org/heigit/bigspatialdata/oshdb/api/db/OSHDBH2.java
@@ -12,7 +12,7 @@ import java.util.Set;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.heigit.bigspatialdata.oshdb.TableNames;
+import org.heigit.bigspatialdata.oshdb.util.TableNames;
 
 /**
  * OSHDB database backend connector to a H2 database.
